Fixing issues on admin name

+++ adminPage.js
@@ -35,48 +35,54 @@ check = () => {
       </div>
     `
   } else {
-    let adminName = []
-    if (localStorage.adminDetails) {
-      adminName = JSON.parse(localStorage.getItem("adminDetails"));
-    }
-    
 
-    Swal.fire({
-      title: `What's Your Name`,
-      input: 'text',
-      inputAttributes: {
-        autocapitalize: 'on'
-      },
-      showCancelButton: false,
-      confirmButtonText: 'Save changes',
-      showLoaderOnConfirm: true,
-    }).then((result) => {
-      if (result.isConfirmed) {
-        Swal.fire(
-          'Good job!',
-           `Wlcome ${result.value}`,
-          'success'
-        )
+    if (!localStorage.adminDetails) {
+      let adminName = []
+      if (localStorage.adminDetails) {
+        adminName = JSON.parse(localStorage.getItem("adminDetails"));
       }
-
-      let theAdminName = {
-        adminFName: result.value
-      }
-      adminName.push(theAdminName);
-      localStorage.setItem("adminDetails", JSON.stringify(adminName));
-      adminName.map((eachAdmin, index) => {
-        localStorage.setItem("currentAdminIndex", index);
+      
+  
+      Swal.fire({
+        title: `What's Your Name`,
+        input: 'text',
+        inputAttributes: {
+          autocapitalize: 'on'
+        },
+        showCancelButton: false,
+        confirmButtonText: 'Save changes',
+        showLoaderOnConfirm: true,
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Good job!',
+             `Wlcome ${result.value}`,
+            'success'
+          )
+        }
+  
+        let theAdminName = {
+          adminFName: result.value
+        }
+        adminName.push(theAdminName);
+        localStorage.setItem("adminDetails", JSON.stringify(adminName));
+        adminName.map((eachAdmin, index) => {
+          localStorage.setItem("currentAdminIndex", index);
+        })
+        window.location.href = `adminPage.html`
       })
-    })
-
-    
-    currentAdminIndex = localStorage.getItem("currentAdminIndex");
-    adminName.map((eachAdmin, index) => {
-      offcanvasWithBothOptionsLabel.innerHTML = `${adminName[index].adminFName}`
-    })
-    adminMain.innerHTML = `
+  
+      
+      
+    } else {
+      adminName = JSON.parse(localStorage.getItem("adminDetails"));
+      currentAdminIndex = localStorage.getItem("currentAdminIndex");
+      adminName.map((eachAdmin, index) => {
+        adminFullName.innerHTML = `${adminName[currentAdminIndex].adminFName}`
+      })
+      adminMain.innerHTML = `
     <div class="shadow admin-btn-container d-flex flex-column w-25 pb-3">
-        <h4 class="text-center mb-3 pt-3">Admin</h4>
+        <h4 class="text-center mb-3 pt-3 text-capitalize">${adminName[currentAdminIndex].adminFName}</h4>
         <div class="d-flex justify-content-center mb-5">
           <img
             src="Images/avatar.png"
@@ -327,6 +333,8 @@ check = () => {
         </div>
       </div>
     `
+    }
+    
   }
 };
